moved tasks to build-system
