+++ build-system/tasks/validate.js
